+++ src/main/java/com/deoxservices/chipsarmorstandmenu/menu/MenuTicker.java
@@ -1,8 +1,9 @@
 package com.deoxservices.chipsarmorstandmenu.menu;
 
-import com.deoxservices.chipsarmorstandmenu.client.ClientProxyGameEvents;
+import com.deoxservices.chipsarmorstandmenu.data.ArmorStandData;
 import com.deoxservices.chipsarmorstandmenu.server.config.ServerConfig;
 import com.deoxservices.chipsarmorstandmenu.utils.Constants;
+
 import net.minecraft.server.level.ServerPlayer;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
@@ -17,7 +18,7 @@ public class MenuTicker {
                 menu.ticksSinceInteraction++;
                 int timeoutTicks = ServerConfig.CONFIG.TIMEOUT_SECONDS.get() * 20;
                 if (menu.ticksSinceInteraction >= timeoutTicks || !menu.stillValid(player)) {
-                    ClientProxyGameEvents.unlockArmorStand(menu.getArmorStand().getId());
+                    ArmorStandData.setArmorStandInUse(menu.getArmorStand(), player.getUUID(), false);
                     player.closeContainer();
                 }
             }